SWupdate Double Copy with Initramfs
This is a multiconfig build to generate an initramfs and a main image.
There are two other build targets, swupdate-image and
swupdate-delta-image which will generate the .swu and zck header.
Additionally, there is support for verified boot to embed a public key
in the u-boot control dtb and sign the FIT image/configurations.

Signed-off-by: Matt Wood <matt.wood@microchip.com>